kubernetes/codes/train/requirements.txt
scikit-learn
pandas
joblib
azure-storage-blob
prometheus-client
pyarrow
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
prometheus-api-client